AirlineFood README.md
This is a program about two people, John and Susan, are travelling on different airlines, but leaving from the same city (Atlanta), and both going non-stop to Bangcock. John is traveling on United Airlines and Susan is on American Airlines. The program will ask them what they each want for their in-flight meals. The choices are beef and broccoli, fish and chips, or spaghetti. It will also ask if they have an dietary restrictions, as John is a vegetarian.